+++ assertion/anonymousfunc/analyzer_test.go
@@ -24,8 +24,8 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.uber.org/goleak"
 	"go.uber.org/nilaway/config"
+	"go.uber.org/nilaway/nilawaytest"
 	"go.uber.org/nilaway/util/analysishelper"
-	"go.uber.org/nilaway/util/nilawaytest"
 	"golang.org/x/tools/go/analysis/analysistest"
 )
